Fix : [WUMO-199] PartyMember 삭제 기능 테스트 오류 수정
- partyParticipant 가 설정되지 않는 테스트 후 clean()에서 deleteById 하는 문제 수정

+++ src/test/java/org/prgrms/wumo/domain/party/api/PartyMemberControllerTest.java
@@ -108,7 +108,10 @@ class PartyMemberControllerTest extends MysqlTestContainer {
 
 	@AfterEach
 	void clean() {
-		partyMemberRepository.deleteById(partyParticipant.getId());
+		if (partyParticipant != null) {
+			partyMemberRepository.deleteById(partyParticipant.getId());
+		}
+		partyParticipant = null;
 		partyMemberRepository.deleteById(partyLeader.getId());
 		partyRepository.deleteById(party.getId());
 		memberRepository.deleteById(leader.getId());
